Rename and add interfaces
