+++ src/omnisharp/utils/getMonoVersion.ts
@@ -1,0 +1,41 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { ChildProcess, spawn } from 'child_process';
+import { IGetMonoVersion } from '../constants/IGetMonoVersion';
+
+export const getMonoVersion: IGetMonoVersion = async (environment: NodeJS.ProcessEnv) => {
+    const versionRegexp = /(\d+\.\d+\.\d+)/;
+
+    return new Promise<string>((resolve, reject) => {
+        let childprocess: ChildProcess;
+        try {
+            childprocess = spawn('mono', ['--version'], { env: environment });
+        }
+        catch (e) {
+            return resolve(undefined);
+        }
+
+        childprocess.on('error', function (err: any) {
+            resolve(undefined);
+        });
+
+        let stdout = '';
+        childprocess.stdout.on('data', (data: NodeBuffer) => {
+            stdout += data.toString();
+        });
+
+        childprocess.stdout.on('close', () => {
+            let match = versionRegexp.exec(stdout);
+
+            if (match && match.length > 1) {
+                resolve(match[1]);
+            }
+            else {
+                resolve(undefined);
+            }
+        });
+    });
+}
